linux-intel: allow up to 16 UARTs
By default, Linux only supports up to 4 8250 UARTs, which is not enough
for many Intel CPUs. Our MB-SMARC-3 makes up to 6 UARTs accessible.

We bump up the maximum to 16 to match
bsp/intel-x86/intel-x86-serial-8250-overrides.cfg in yocto-kernel-cache.
In a future BSP version, we might be able to switch to the Yocto intel-x86
default config altogether, but for now we keep the changes small.

A side effect of having a large number of spare UART numbers is that the
driver will always reserve ttyS0 to ttyS3 for ioport-based instances,
while MMIO-based UARTs start at ttyS4, causing renumbering on some of
our boards.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>