linux-yocto: add TQMx86 patches and configuration
The TQMx86 default kernel will be based on linux-yocto instead of
linux-intel from now on to be closer to the well-maintained linux-stable
versions. bbappends are provided for both linux-yocto and linux-yocto-rt.

This adds a number of patches for TQMx86 COMs:

- Backports to work around a pinctrl-intel issue with old TQMxE40x BIOS
  versions
- Backports and additional patches to fix m_can issues on TQMxE40x
- The latest verson of the TQMx86 MFD and GPIO patches

The KMACHINE is changed to intel-x86-64, which provides better defaults
than intel-corei7-64, allowing us to drop elkhart-lake.cfg and serial.cfg.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>